Port over some drf-yasg functionality, attempting to add explanatory comments
